requirements.txt
streamlit
pandas
pandas_datareade
datetim
yfinance

prophet
matplotlib

